added correct marshalling and unmarshalling
